Updating SystemD unit files to make use of SystemD's feature to manage runtime directories when possible.
When the service does not run as roo user and if the desired runtime directory is under /run or /var/run (which is a usually a symlink to /run), then the explicit creation of the directory in a ExecStartPre setting should be replaced by the RuntimeDirectory setting which will have expected outcome.
Otherwise, if the runtime directory is located somewhere else, the previous approach is preserved.

See: https://github.com/echocat/puppet-redis/issues/92