Add 2 new Redis configuration options related to AOF
Add configuration for:
aof-load-truncated - default is 'yes' in Redis
aof-use-rdb-preamble - default 'no' in Redis

But both configuration options default to 'undef' in the manifest
in order to omit the options from the generated configuration file
if not passed to the class. This way older versions of Redis managed by
this module will not get unknown configuration options.