Add 2 new Redis configuration options related to Replication
Add configuration for:
repl-diskless-sync - default is 'no' in Redis
repl-diskless-sync-delay - default is '5' in Redis

But both configuration options default to 'undef' in the manifest
in order to omit the options from the generated configuration file
if not passed to the class. This way older versions of Redis managed by
this module will not get unknown configuration options.